Move weak_lang_items checking to librustc_passes.
