+++ src/librustc_passes/weak_lang_items.rs
@@ -1,21 +1,18 @@
 //! Validity checking for weak lang items
 
-use crate::middle::lang_items;
-use crate::session::config;
+use rustc::middle::lang_items;
+use rustc::middle::weak_lang_items::whitelisted;
+use rustc::session::config;
 
-use crate::hir::map::Map;
-use crate::ty::TyCtxt;
+use rustc::hir::map::Map;
+use rustc::ty::TyCtxt;
 use rustc_data_structures::fx::FxHashSet;
 use rustc_errors::struct_span_err;
 use rustc_hir as hir;
-use rustc_hir::def_id::DefId;
 use rustc_hir::intravisit::{self, NestedVisitorMap, Visitor};
 use rustc_lang_items::weak_lang_items::WEAK_ITEMS_REFS;
 use rustc_span::symbol::Symbol;
 use rustc_span::Span;
-use rustc_target::spec::PanicStrategy;
-
-pub use rustc_lang_items::weak_lang_items::link_name;
 
 struct Context<'a, 'tcx> {
     tcx: TyCtxt<'tcx>,
@@ -40,24 +37,6 @@ pub fn check_crate<'tcx>(tcx: TyCtxt<'tcx>, items: &mut lang_items::LanguageItem
         tcx.hir().krate().visit_all_item_likes(&mut cx.as_deep_visitor());
     }
     verify(tcx, items);
-}
-
-/// Returns `true` if the specified `lang_item` doesn't actually need to be
-/// present for this compilation.
-///
-/// Not all lang items are always required for each compilation, particularly in
-/// the case of panic=abort. In these situations some lang items are injected by
-/// crates and don't actually need to be defined in libstd.
-pub fn whitelisted(tcx: TyCtxt<'_>, lang_item: lang_items::LangItem) -> bool {
-    // If we're not compiling with unwinding, we won't actually need these
-    // symbols. Other panic runtimes ensure that the relevant symbols are
-    // available to link things together, but they're never exercised.
-    if tcx.sess.panic_strategy() != PanicStrategy::Unwind {
-        return lang_item == lang_items::EhPersonalityLangItem
-            || lang_item == lang_items::EhUnwindResumeLangItem;
-    }
-
-    false
 }
 
 fn verify<'tcx>(tcx: TyCtxt<'tcx>, items: &lang_items::LanguageItems) {
@@ -120,11 +99,5 @@ impl<'a, 'tcx, 'v> Visitor<'v> for Context<'a, 'tcx> {
             self.register(lang_item, i.span);
         }
         intravisit::walk_foreign_item(self, i)
-    }
-}
-
-impl<'tcx> TyCtxt<'tcx> {
-    pub fn is_weak_lang_item(&self, item_def_id: DefId) -> bool {
-        self.lang_items().is_weak_lang_item(item_def_id)
     }
 }
